refactor: separate process and dialog writing
